requirements.txt
requests==2.32.3
Flask==3.1.0
python-dotenv==1.1.0
gunicorn==23.0.0
azure-data-tables==12.6.0
azure-storage-blob==12.25.1
langchain==0.3.23
langchain-openai==0.3.11
langchain-deepseek==0.1.3
langchain-community==0.3.21
langgraph==0.3.21
xmltodict==0.14.2
pydantic==2.11.1
duckduckgo-search==8.0.1
playwright
lxml
beautifulsoup4